Every JSONRPC JSON structure deserves its own file
